requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.6.0
idna==3.1
idna-ssl==1.1.0
multidict==5.1.0
SQLAlchemy==1.3.22
websockets==8.1
typing-extensions==3.7.4.3
yarl==1.6.3
